[REF]: using python modules
The index page is now generated using a module containing the
generating node function. This should be more efficient since the
parser does not have to recompile all the code it had written. Instead
now it loads a pre-compiled python module and runs it. Additionally,
the index page will be generated from any working directory thanks to
the keyword `__DIR__` which is the path to the directory containing the
`index.pyhtml` file.