readthedocs theme
lexor is still in its developing stages and cannot yet take onto
creating its own documentation pages without first creating a program
that would use lexor to manage all the files. For this reason this
repository will use sphinx and the readthedocs theme to create pages.
The readthedocs theme will be modified to give the lexor styles
documentation some character of its own.

The settings have been changed from the original source in order to
start testing different styles. For now I set the headers to be of a
ridiculous green color to make sure everything was set up correctly.